Moved audio flag commands to new tree view and all related code into new file
